added song delete confirmation and made window resize properly

+++ myTunes/MainWindow.xaml.cs
@@ -133,7 +133,11 @@ namespace myTunes
                 {
                     var dataRow = data.Row;
                     int songId = (int)dataRow["Id"];
-                    musicRepo.DeleteSong(songId);
+                    MessageBoxResult result = MessageBox.Show($"Are you sure you want to delete \"{dataRow["Title"].ToString()}\"?", "Confirmation", MessageBoxButton.YesNo, MessageBoxImage.Question);
+                    if (result == MessageBoxResult.Yes)
+                    {
+                        musicRepo.DeleteSong(songId);
+                    }
                 }
             }
         }
